requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
crispy-bootstrap4==2024.10
Django==5.1.3
django-bootstrap4==24.3
django-crispy-forms==2.3
django-taggit==5.0.1
djangorestframework==3.14.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
openai==0.27.6
pillow==11.0.0
propcache==0.2.0
pytz==2024.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
